backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.104.1",
    "ibm-watsonx-ai>=1.4.7",
    "pydantic==2.5.0",
    "python-multipart==0.0.6",
    "uvicorn[standard]==0.24.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "python-dotenv>=1.0.0",
]
